Add files via upload
Initial commit. This is a rough outline of the site. Footer still needs to be done, navbar needs social icons, another section may be added, media queries need to be added for mobile view(along with adjusting typography sizes) and a scrollspy will be added to allow the user to see which part of the site they are on.